+++ app/controllers/search_page_controller.rb
@@ -1,0 +1,4 @@
+class SearchPageController < CmsController
+  def index
+  end
+end