+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # v1.1.0
-  * The search has been moved into a new search_page component. It adds a new page type SearchPage as well as a separate cell for the search box.
-    See `rails generate cms:component:search_page --help` for more details.
+  * Search support moved into its own generator and got cleaned up and simplified. See
+    `rails generate cms:component:search --help` for more details.
   * Edit-Marker no longer are included by Infopark Kickstarter, as it became a default feature of
     of Infopark RailsConnector. (Thanks @tritz)
   * All rake tasks now use RestClient instead of curl, for better compatibility and consistent use